+ [x] Don't zero the token internally when emitting it on `.next()`
+ [x] Use `unchecked_read`s instead of passing array pointers for multibyte reads.
+ [ ] Test deliberate jump tables (`[fn(&mut Lexer); N]`).
+ [ ] Happy path for `max_read` bytes.

So far so good:
```
test identifiers                       ... bench:         591 ns/iter (+/- 23) = 1318 MB/s
test keywords_operators_and_punctators ... bench:       2,122 ns/iter (+/- 42) = 1004 MB/s
test strings                           ... bench:         593 ns/iter (+/- 43) = 1468 MB/s
```